split sectionD part
